fix casting bug for newer GCC version

+++ src/worker/key_values_rdd.h
@@ -75,7 +75,7 @@ class KeyValuesRDD: public RDD {
     std::vector<msgpack::sbuffer> buffers(num_partitions);
     Pack(buffers);
     for (int i = 0; i < num_partitions; ++i) {
-      block_mgr.PutBlock(i, buffers[i].size(), std::unique_ptr<char[]>(buffers[i].release()));
+      block_mgr.PutBlock(i, static_cast<int32_t>(buffers[i].size()), std::unique_ptr<char[]>(buffers[i].release()));
     }
     key_values_.clear();
   }
